move outputaccumulator.py
