+++ alphadia/outputtransform/search_plan_output.py
@@ -11,15 +11,15 @@ from alphadia import grouping, libtransform, utils
 from alphadia.constants.keys import ConfigKeys
 from alphadia.constants.settings import FIGURES_FOLDER_NAME
 from alphadia.exceptions import NoPsmFoundError
-from alphadia.outputaccumulator import (
-    AccumulationBroadcaster,
-    TransferLearningAccumulator,
-)
 from alphadia.outputtransform.df_builders import (
     build_run_internal_df,
     build_run_stat_df,
     log_stat_df,
     transfer_library_stat_df,
+)
+from alphadia.outputtransform.outputaccumulator import (
+    AccumulationBroadcaster,
+    TransferLearningAccumulator,
 )
 from alphadia.outputtransform.protein_fdr import perform_protein_fdr
 from alphadia.outputtransform.quant_builder import QuantBuilder
