+++ tests/unit_tests/test_outputaccumulator.py
@@ -8,7 +8,7 @@ from alphabase.spectral_library.base import SpecLibBase
 from conftest import mock_fragment_df, mock_precursor_df
 
 from alphadia.constants.keys import SearchStepFiles
-from alphadia.outputaccumulator import ms2_quality_control
+from alphadia.outputtransform.outputaccumulator import ms2_quality_control
 from alphadia.outputtransform.search_plan_output import SearchPlanOutput
 from alphadia.workflow.base import QUANT_FOLDER_NAME
